+++ resteasy-client-api/src/main/java/org/jboss/resteasy/client/jaxrs/PublisherRxInvoker.java
@@ -1,4 +1,4 @@
-package org.jboss.resteasy.client.jaxrs.internal;
+package org.jboss.resteasy.client.jaxrs;
 
 import javax.ws.rs.client.Entity;
 import javax.ws.rs.client.RxInvoker;
