+++ resteasy-reactor/src/main/java/org/jboss/resteasy/reactor/MonoRxInvoker.java
@@ -4,7 +4,7 @@ import javax.ws.rs.client.Entity;
 import javax.ws.rs.core.GenericType;
 import javax.ws.rs.core.Response;
 
-import org.jboss.resteasy.client.jaxrs.internal.PublisherRxInvoker;
+import org.jboss.resteasy.client.jaxrs.PublisherRxInvoker;
 import reactor.core.publisher.Mono;
 
 public interface MonoRxInvoker extends PublisherRxInvoker
